backend/requirements.txt
scikit-learn
transformers
torch
motor
json-logging
aiofiles
aiohappyeyeballs
aiohttp
aiosignal
annotated-types
anthropic
anyio
asgiref
async-timeout
attrs
Authlib
blinker
cachetools
certifi
cffi
charset-normalizer
click
cryptography
dataclasses-json
distro
dnspython
exceptiongroup
filelock
Flask
Flask-Cors
frozenlist
fsspec
google-ai-generativelanguage
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-generativeai
google_search_results
googleapis-common-protos
greenlet
grpcio
grpcio-health-checking
grpcio-status
grpcio-tools
h11
httpcore
httplib2
httpx
huggingface-hub
idna
importlib_metadata
itsdangerous
Jinja2
jiter
joblib
json-logging
jsonpatch
jsonpointer
langchain
langchain-community
langchain-core
langchain-openai
langchain-text-splitters
langsmith
MarkupSafe
marshmallow
motor
mpmath
multidict
mypy-extensions
networkx
numpy
openai
orjson
packaging
proto-plus
protobuf
pyasn1
pyasn1_modules
pycparser
pydantic
pydantic_core
pymongo
pyparsing
python-dotenv
PyYAML
regex
requests
rsa
safetensors
scikit-learn
scipy
sniffio
SQLAlchemy
sympy
tenacity
threadpoolctl
tiktoken
tokenizers
torch
tqdm
transformers
typing-inspect
typing_extensions
uritemplate
urllib3
validators
weaviate-client>=4.0.0
Werkzeug
yarl
zipp
